requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
mysql-connector-python==8.0.32
Pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
